fixes for ArcoLinuxa and kitty ssh

+++ .scripts/install_init.zsh
@@ -13,25 +13,27 @@ initInstall "install_init"
 #####################
 
 # Config pacman
-sed 's/^#Color$/Color/g' </etc/pacman.conf >pacman.conf
-sudo mv pacman.conf /etc/
-sed 's/^.*ParallelDownloads.*$/ParallelDownloads = 5/g' </etc/pacman.conf >pacman.conf
-sudo mv pacman.conf /etc/
-errorCheck "pacman config"
+if [ -f /etc/pacman.conf ]; then
+	sed 's/^#Color$/Color/g' </etc/pacman.conf >pacman.conf
+	sudo mv pacman.conf /etc/
+	sed 's/^.*ParallelDownloads.*$/ParallelDownloads = 5/g' </etc/pacman.conf >pacman.conf
+	sudo mv pacman.conf /etc/
+	errorCheck "pacman config"
+fi
 
 # Config pamac
-sed 's/^.*ILoveCandy$/ILoveCandy/g' </etc/pamac.conf >pamac.conf
-sudo mv pamac.conf /etc/
-sed 's/^.*EnableAUR$/EnableAUR/g' </etc/pamac.conf >pamac.conf
-sudo mv pamac.conf /etc/
-sed 's/^.*KeepBuiltPkgs$/KeepBuiltPkgs/g' </etc/pamac.conf >pamac.conf
-sudo mv pamac.conf /etc/
-errorCheck "pamac config"
+if [ -f /etc/pamac.conf ]; then
+	sed 's/^.*ILoveCandy$/ILoveCandy/g' </etc/pamac.conf >pamac.conf
+	sudo mv pamac.conf /etc/
+	sed 's/^.*EnableAUR$/EnableAUR/g' </etc/pamac.conf >pamac.conf
+	sudo mv pamac.conf /etc/
+	sed 's/^.*KeepBuiltPkgs$/KeepBuiltPkgs/g' </etc/pamac.conf >pamac.conf
+	sudo mv pamac.conf /etc/
+	errorCheck "pamac config"
+fi
 
-sudo pacman -S --noconfirm --needed git base-devel colorgcc go ruby rust yay
+sudo pacman -S --noconfirm --needed git base-devel colorgcc go ruby rust
 errorCheck "installation base-devel"
-
-git submodule update --init --recursive
 
 # Prompt installieren
 inst ttf-meslo-nerd-font-powerlevel10k
